Adding class for discrete trajectory
