+++ spline_slam/trajectory/__init__.py
@@ -1,0 +1,1 @@
+from .discrete_trajectory import DiscreteTrajectory
